Move security-related functions to own module.
